Sub halo looping implemented Subhalo branching has now been implemented. This is done by using a structured array to save all subhalo properties as a single attribute within its host halo class.
An extra ipython notebook has been added to showcase this. Currently there is no
output as I am not sure I have the data for the star masses etc. I will look into
this soon.

A few things will be removed when Will updates the file format. The main thing
is the try-except statement which is triggered when there are no subhalos
present.

All docstrings are up to date too!