Large changes. C routines are now called and used.
The model infall and cooling routines have now been implemented. They are called each snapshot although I'm not entirely sure if they are called
at the correct point. Everything seems to be working well and as expected. Merging of L-galaxies with Py-galaxies was a pain though. In the main
driver notebook (Deleted two redundant ones) you can see that in the import cell I've had to include an extra path to the c codes etc. Another
interesting feature is that the relative filepathing from C code to txt file data tables is from the DRIVER NOTEBOOk and not from the c code.

Everything working well though. Major updates to documentation needed. Next step is to make several plots from the new functions and also
the virial temperature too as that was all implemented today also.